% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_hyperparameters.R
\name{mixture}
\alias{mixture}
\title{Make a Mixture of Hyperparameter Distributions}
\usage{
mixture(..., weights)
}
\arguments{
\item{...}{Distributions created by \code{\link{make_discrete}} or \code{\link{make_range}}.}

\item{weights}{Numeric vector of non-negative weights corresponding to each distribution in \code{...}.
The length of \code{weights} must match the number of distributions provided.
If weights do not sum to 1, they will be normalized internally.}
}
\value{
An object of class \code{enfold_mixture}.
}
\description{
Takes an arbitrary amount of sample spaces made via
\code{\link{make_discrete}} and \code{\link{make_range}} and combines them
into a single mixture distrubution. The \code{weights} argument
controls the relative sampling frequency of each component distribution.
}
\examples{
# Draw a 0 half the time, and a number between 0 and 10 half the time
mixture_dist <- mixture(
  make_discrete(0),
  make_range(0, 10),
  weights = c(0.5, 0.5)
)
}
\seealso{
\code{\link{specify_hyperparameters}}, \code{\link{make_range}}, \code{\link{make_discrete}},
\code{\link{draw}}
}
