% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_hyperparameters.R
\name{sample_log_uniform}
\alias{sample_log_uniform}
\title{Log-Uniform Sampler for Continuous Hyperparameters}
\usage{
sample_log_uniform()
}
\value{
A function of n that generates n random values log-uniformly distributed between the specified min and max of the range.
}
\description{
Provides a sampling function that generates log-uniform random values.
}
