% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_hyperparameters.R
\name{sample_uniform}
\alias{sample_uniform}
\title{Uniform Sampler for Continuous Hyperparameters}
\usage{
sample_uniform()
}
\value{
A function of n that generates n random values uniformly distributed between the specified min and max of the range.
}
\description{
Provides a sampling function that generates uniform random values.
Is the default sampling method for continuous hyperparameters defined by \code{make_range}.
}
